Comando /config, dropando user_config.default_sheet e várias correções em outros sistemas

+++ src/resources/utils/action-handler.ts
@@ -1,34 +1,48 @@
 import logger from '../../configs/logger';
 
 class ActionHandler {
-    private actions: Map<string, { createdAt: Date; expireAt: Date; singleUse: boolean; run: Function }> = new Map();
+    private actions: Map<
+        string,
+        { createdAt: Date; expireAt: Date; singleUse: boolean; respondOnlyToUserId?: string; run: Function }
+    > = new Map();
     constructor() {
-        setInterval(() => {
-            this.clearExpiredActions();
-        }, 10 * 60 * 1000);
+        setInterval(
+            () => {
+                this.clearExpiredActions();
+            },
+            10 * 60 * 1000
+        );
     }
 
-    registerAction(id: string, data: { action: Function; expireAt?: Date; singleUse?: boolean }): void {
+    registerAction(
+        id: string,
+        data: { action: Function; expireAt?: Date; singleUse?: boolean; respondOnlyToUserId?: string }
+    ): void {
         const now = new Date();
         this.actions.set(id, {
             createdAt: now,
             expireAt: data.expireAt ? data.expireAt : new Date(now.getTime() + 18 * 60 * 1000),
             singleUse: data.singleUse ? true : false,
-            run: data.action
+            run: data.action,
+            respondOnlyToUserId: data.respondOnlyToUserId
         });
     }
 
-    executeAction(id: string, ...args: any[]): unknown | void {
+    executeAction(id: string, userId: string, ...args: any[]): unknown | void {
         const action = this.actions.get(id);
 
         if (action) {
             if (action.expireAt > new Date()) {
                 try {
-                    if (action.singleUse) {
-                        this.deleteAction(id);
-                    }
+                    if (action.respondOnlyToUserId && action.respondOnlyToUserId !== userId) {
+                        return;
+                    } else {
+                        if (action.singleUse) {
+                            this.deleteAction(id);
+                        }
 
-                    return action.run(...args);
+                        return action.run(...args);
+                    }
                 } catch (err) {
                     logger.logText('WARN', `Error executing action "${id}": ${err}`);
                 }
